:truck: move cmd util to internal/utils
