+++ cmd/generate.go
@@ -3,13 +3,10 @@ package cmd
 
 import (
 	"bytes"
-	"fmt"
 	"github.com/nikhilhenry/xframe/pkg/frame"
 	"github.com/spf13/cobra"
 	"image"
 	"os"
-	"path"
-	"strings"
 )
 
 // generateCmd represents the frame command
@@ -49,24 +46,6 @@ using official Apple device bezels.`,
 		}
 		return nil
 	},
-}
-
-func getFilePath(imageFilePath string, rawPath string) (filePath string) {
-	if rawPath != "." {
-		if strings.HasSuffix(rawPath, ".png") {
-			filePath = rawPath
-			return
-		}
-		filePath = rawPath + ".png"
-		return
-	}
-	cleanedPath := path.Clean(rawPath)
-	filePathElements := strings.SplitAfter(imageFilePath, "/")
-	fmt.Println(filePathElements)
-	fileNameWithExtension := filePathElements[len(filePathElements)-1]
-	fileName := strings.Split(fileNameWithExtension, ".")[0]
-	filePath = fmt.Sprintf("%s/%s-framed.png", cleanedPath, fileName)
-	return
 }
 
 func init() {
